refactor: refine type
