+++ packages/bezier-react/src/components/LegacyStack/LegacyHStack/LegacyHStack.tsx
@@ -1,5 +1,4 @@
 import React, { forwardRef } from 'react'
-import type { Ref } from 'react'
 
 import { LegacyStack } from '~/src/components/LegacyStack/LegacyStack'
 
@@ -8,9 +7,9 @@ import type LegacyHStackProps from './LegacyHStack.types'
 /**
  * A container for horizontal flex layout.
  */
-export const LegacyHStack = forwardRef(function HStack(
-  props: LegacyHStackProps,
-  forwardedRef: Ref<HTMLElement>,
+export const LegacyHStack = forwardRef<HTMLElement, LegacyHStackProps>(function HStack(
+  props,
+  forwardedRef,
 ) {
   return (<LegacyStack ref={forwardedRef} direction="horizontal" {...props} />)
 })
