+++ packages/bezier-react/src/components/LegacyStack/LegacyHStack/LegacyHStack.types.ts
@@ -1,3 +1,3 @@
-import type { LegacyStackProps } from '~/src/components/LegacyStack/LegacyStack'
+import type { LegacyStackProps } from '~/src/components/LegacyStack'
 
 export default interface LegacyHStackProps extends Omit<LegacyStackProps, 'direction'> {}
