+++ packages/bezier-react/src/components/LegacyStack/LegacySpacer/LegacySpacer.tsx
@@ -1,13 +1,12 @@
 import React, { forwardRef } from 'react'
-import type { Ref } from 'react'
 
 import { LegacyStackItem } from '~/src/components/LegacyStack/LegacyStackItem'
 
 import type LegacySpacerProps from './LegacySpacer.types'
 
-export const LegacySpacer = forwardRef(function Spacer(
-  props: LegacySpacerProps,
-  forwardedRef: Ref<HTMLElement>,
+export const LegacySpacer = forwardRef<HTMLElement, LegacySpacerProps>(function Spacer(
+  props,
+  forwardedRef,
 ) {
   return (
     <LegacyStackItem
