+++ packages/bezier-react/src/components/LegacyStack/LegacyVStack/LegacyVStack.tsx
@@ -1,5 +1,4 @@
 import React, { forwardRef } from 'react'
-import type { Ref } from 'react'
 
 import { LegacyStack } from '~/src/components/LegacyStack/LegacyStack'
 
@@ -8,9 +7,9 @@ import type LegacyVStackProps from './LegacyVStack.types'
 /**
  * A container for vertical flex layout.
  */
-export const LegacyVStack = forwardRef(function VStack(
-  props: LegacyVStackProps,
-  forwardedRef: Ref<HTMLElement>,
+export const LegacyVStack = forwardRef<HTMLElement, LegacyVStackProps>(function VStack(
+  props,
+  forwardedRef,
 ) {
   return (<LegacyStack ref={forwardedRef} direction="vertical" {...props} />)
 })
